% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxProfileItemT.R
\name{boxProfileItemT}
\alias{boxProfileItemT}
\title{Custom Boxprofile Item}
\usage{
boxProfileItemT(title = NULL, description = NULL)
}
\arguments{
\item{title}{Title}

\item{description}{Description}
}
\value{

}
\description{
This function generates a custom Boxprofile Item. Built from shinydashboardPlus boxProfileItem.
}
\details{
Requires:
}
\examples{

}
