% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxProfileT.R
\name{boxProfileT}
\alias{boxProfileT}
\title{Custom Boxprofile}
\usage{
boxProfileT(..., title = NULL, subtitle = NULL)
}
\arguments{
\item{title}{Title}

\item{subtitle}{Subtitle}
}
\value{

}
\description{
This function generates a custom Boxprofile. Built from shinydashboardPlus boxProfile.
}
\details{
Requires:
}
\examples{

}
