% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdownMenuT.R
\name{dropdownMenuT}
\alias{dropdownMenuT}
\title{Custom Dropdown Menu}
\usage{
dropdownMenuT(..., type = c("notifications"), profileName = NULL, .list = NULL)
}
\arguments{
\item{type}{Type}

\item{profileName}{Profile Name}
}
\value{

}
\description{
This function generates a custom Boxprofile Item. Built from shinydashboard dropdownMenu.
}
\details{
Requires:
}
\examples{

}
