% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBoxT.R
\name{infoBoxT}
\alias{infoBoxT}
\title{Custom Infobox}
\usage{
infoBoxT(
  title,
  value = NULL,
  subtitle = NULL,
  caption = NULL,
  icon = shiny::icon("bar-chart"),
  color = "#ffffff",
  width = 4,
  href = NULL,
  fill = FALSE
)
}
\arguments{
\item{title}{Title}

\item{value}{Value}

\item{subtitle}{Subtitle}

\item{caption}{Caption}
}
\value{

}
\description{
This function generates a custom Infobox. Built from shinydashboard infoBox.
}
\details{
Requires:
}
\examples{

}
